STUB: save preferred braces for macro 2.0
It will help to suggest proper braces during completion
