+++ src/main/kotlin/org/rust/lang/core/stubs/StubImplementations.kt
@@ -67,7 +67,7 @@ class RsFileStub(
     override fun getType() = Type
 
     object Type : IStubFileElementType<RsFileStub>(RsLanguage) {
-        private const val STUB_VERSION = 216
+        private const val STUB_VERSION = 217
 
         // Bump this number if Stub structure changes
         override fun getStubVersion(): Int = RustParserDefinition.PARSER_VERSION + STUB_VERSION
@@ -1538,6 +1538,7 @@ class RsMacro2Stub(
     override val name: String?,
     val macroBody: String,
     val bodyHash: HashCode,
+    val preferredBraces: MacroBraces,
     override val flags: Int
 ) : RsAttributeOwnerStubBase<RsMacro2>(parent, elementType),
     RsNamedStub {
@@ -1553,6 +1554,7 @@ class RsMacro2Stub(
                 dataStream.readNameAsString(),
                 dataStream.readUTFFast(),
                 dataStream.readHashCode(),
+                dataStream.readEnum(),
                 dataStream.readUnsignedByte()
             )
 
@@ -1561,6 +1563,7 @@ class RsMacro2Stub(
                 writeName(stub.name)
                 writeUTFFast(stub.macroBody)
                 writeHashCode(stub.bodyHash)
+                writeEnum(stub.preferredBraces)
                 writeByte(stub.flags)
             }
 
@@ -1572,7 +1575,8 @@ class RsMacro2Stub(
             flags = BitUtil.set(flags, HAS_RUSTC_BUILTIN_MACRO, MACRO2_HAS_RUSTC_BUILTIN_MACRO_PROP.getDuringIndexing(psi))
             val body = psi.prepareMacroBody()
             val bodyHash = HashCode.compute(body)
-            return RsMacro2Stub(parentStub, this, psi.name, body, bodyHash, flags)
+            val preferredBraces = psi.preferredBraces
+            return RsMacro2Stub(parentStub, this, psi.name, body, bodyHash, preferredBraces, flags)
         }
 
         override fun indexStub(stub: RsMacro2Stub, sink: IndexSink) = sink.indexMacroDef(stub)
